refactor : OperatorResolver Repository에서 Resolver로 의미가 더 잘 전달되도록 클래스 이름 변경
