+++ src/main/java/calculator/operator/OperatorResolver.java
@@ -2,13 +2,13 @@ package calculator.operator;
 
 import java.util.List;
 
-public class OperatorRepository {
+public class OperatorResolver {
 
     public static final String INPUT_OPERATOR_EXCEPTION = "입력하신 연산자가 사칙 연산 연산자가 아닙니다.";
 
     private final List<OperatorStrategy> operators;
 
-    public OperatorRepository(List<OperatorStrategy> operators) {
+    public OperatorResolver(List<OperatorStrategy> operators) {
         this.operators = operators;
     }
